feat: modularize ticket updating further
