+++ services/api-main/src/app/tickets/domain/services/ticket-comment.service.ts
@@ -7,27 +7,24 @@ import { TicketNotFoundError } from '../errors/TicketNotFound';
 import { CommentNotFoundError } from '../errors/CommentNotFound';
 import { CannotUpdateOthersCommentsError } from '../errors/CannotUpdateOthersComments';
 import { User } from 'src/app/users/domain/entities/user.entity';
+import { Ticket } from '../entities/ticket.entity';
+import { TicketsRedactionService } from './ticket-redacation.service';
 
 @Injectable()
 export class TicketsCommentService extends BaseService {
   constructor(
     @InjectMapper() private readonly mapper: Mapper,
+    private ticketRedactionService: TicketsRedactionService,
     private ticketsRepository: TicketsRepository,
   ) {
     super();
   }
 
-  async updateComment(id: string, user: User, commentId: string, body: string) {
-    const ticket = await this.ticketsRepository.findById(id);
-
-    if (!ticket) {
-      throw new TicketNotFoundError(id);
-    }
-
+  private findAndProtectComment(ticket: Ticket, user: User, commentId: string) {
     const isOwner = ticket.createdBy.id === user.id;
 
     if (user.isCustomer() && !isOwner) {
-      throw new TicketNotFoundError(id);
+      throw new TicketNotFoundError(ticket.id);
     }
 
     const comment = ticket.comments.find((c) => c.commentId === commentId);
@@ -40,9 +37,17 @@ export class TicketsCommentService extends BaseService {
       throw new CannotUpdateOthersCommentsError();
     }
 
-    if (comment.body === body) {
-      return ticket;
+    return comment;
+  }
+
+  async updateComment(id: string, user: User, commentId: string, body: string) {
+    const ticket = await this.ticketsRepository.findById(id);
+
+    if (!ticket) {
+      throw new TicketNotFoundError(id);
     }
+
+    const comment = this.findAndProtectComment(ticket, user, commentId);
 
     comment.body = body;
 
@@ -52,6 +57,7 @@ export class TicketsCommentService extends BaseService {
       user,
     );
 
+    this.ticketRedactionService.prepareTicketResponse(ticket, user);
     return updatedTicket;
   }
